fix: stabilize investment flows

+++ src/main/java/com/trabalho/gestao_acoes/domains/dtos/PosicaoDTO.java
@@ -1,6 +1,7 @@
 package com.trabalho.gestao_acoes.domains.dtos;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 public class PosicaoDTO implements Serializable {
     private static final long serialVersionUID = 1L;
@@ -8,12 +9,12 @@ public class PosicaoDTO implements Serializable {
     private String ticker;
     private String corretora;
     private Integer quantidade;
-    private Double precoMedio;
+    private BigDecimal precoMedio;
     private String moeda; // <--- NOVO CAMPO
 
     public PosicaoDTO() {}
 
-    public PosicaoDTO(String ticker, String corretora, Integer quantidade, Double precoMedio, String moeda) {
+    public PosicaoDTO(String ticker, String corretora, Integer quantidade, BigDecimal precoMedio, String moeda) {
         this.ticker = ticker;
         this.corretora = corretora;
         this.quantidade = quantidade;
@@ -30,8 +31,8 @@ public class PosicaoDTO implements Serializable {
     public Integer getQuantidade() { return quantidade; }
     public void setQuantidade(Integer quantidade) { this.quantidade = quantidade; }
 
-    public Double getPrecoMedio() { return precoMedio; }
-    public void setPrecoMedio(Double precoMedio) { this.precoMedio = precoMedio; }
+    public BigDecimal getPrecoMedio() { return precoMedio; }
+    public void setPrecoMedio(BigDecimal precoMedio) { this.precoMedio = precoMedio; }
 
     public String getMoeda() { return moeda; }
     public void setMoeda(String moeda) { this.moeda = moeda; }
