+++ src/main/java/com/trabalho/gestao_acoes/services/ports/CotacaoBolsa.java
@@ -1,20 +1,22 @@
 package com.trabalho.gestao_acoes.services.ports;
 
+import java.math.BigDecimal;
+
 public class CotacaoBolsa {
 
-    private Double precoAtual;
+    private BigDecimal precoAtual;
     private String moeda; // "BRL" ou "USD"
 
     public CotacaoBolsa() {
     }
 
-    public CotacaoBolsa(Double precoAtual, String moeda) {
+    public CotacaoBolsa(BigDecimal precoAtual, String moeda) {
         this.precoAtual = precoAtual;
         this.moeda = moeda;
     }
 
-    public Double getPrecoAtual() { return precoAtual; }
-    public void setPrecoAtual(Double precoAtual) { this.precoAtual = precoAtual; }
+    public BigDecimal getPrecoAtual() { return precoAtual; }
+    public void setPrecoAtual(BigDecimal precoAtual) { this.precoAtual = precoAtual; }
 
     public String getMoeda() { return moeda; }
     public void setMoeda(String moeda) { this.moeda = moeda; }
